CoreFuzzer/requirements.txt
python-dotenv
pymongo
PyYAML
